drf/cars/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
aps==0.0.0
APScheduler==3.10.4
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
bpython==0.24
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.1
curtsies==0.4.2
cwcwidth==0.1.9
debugpy==1.8.0
decorator==5.1.1
Django==4.2
django-apscheduler==0.6.2
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.5
django-task==2.0.6
djangorestframework==3.14.0
drf-yasg==1.21.7
environs==10.0.0
executing==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
inflection==0.5.1
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
lxml==5.1.0
magic-filter==1.0.12
Markdown==3.5.2
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.5
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyTelegramBotAPI==4.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
rpds-py==0.17.1
schedule==1.2.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
swagger-spec-validator==3.0.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.13
yarl==1.9.4
